Revert "Issue 12/persist genre musicalbum"
