+++ data/DB/books.json
@@ -1,3 +1,20 @@
 [
-
+  {
+    "id": 542,
+    "publisher": "dfhsh",
+    "cover_state": "sdfdsg",
+    "publish_date": "2000-05-05",
+    "genre": "dhfdfhf",
+    "author": "ss",
+    "label": "#<Label:0x0000026ce0ed78b8>"
+  },
+  {
+    "id": 350,
+    "publisher": "dgsdfg",
+    "cover_state": "sdfsgs",
+    "publish_date": "2000-05-05",
+    "genre": "sfgsgd",
+    "author": "dsgsd",
+    "label": "#<Label:0x000001ad68e630e0>"
+  }
 ]
