+++ data/DB/labels.json
@@ -1,3 +1,42 @@
 [
-
+  {
+    "id": 21,
+    "title": "222222222",
+    "color": "44444444444"
+  },
+  {
+    "id": 907,
+    "title": "66666666666",
+    "color": "77777777777"
+  },
+  {
+    "id": 892,
+    "title": "sdfgf",
+    "color": ""
+  },
+  {
+    "id": 452,
+    "title": "dfghd",
+    "color": "dfghgf"
+  },
+  {
+    "id": 558,
+    "title": "sdfd",
+    "color": "sdfgfd"
+  },
+  {
+    "id": 567,
+    "title": "sdgfd",
+    "color": "sfg"
+  },
+  {
+    "id": 987,
+    "title": "sdgdsg",
+    "color": "sdgds"
+  },
+  {
+    "id": 801,
+    "title": "sdgfg",
+    "color": "sdfg"
+  }
 ]
