+++ data/DB/music_album.json
@@ -1,22 +1,11 @@
 [
   {
-    "id": 828,
+    "id": 814,
     "genre": {
-      "id": 411,
+      "id": 586,
       "genre": "a",
       "items": [
-        {
-          "item_id": 828,
-          "item_class": "MusicAlbum"
-        },
-        {
-          "item_id": 429,
-          "item_class": "MusicAlbum"
-        },
-        {
-          "item_id": 335,
-          "item_class": "MusicAlbum"
-        }
+        "#<MusicAlbum:0x000001c97df25948>"
       ]
     },
     "author": "a",
@@ -24,57 +13,5 @@
     "publish_date": "2022-02-02",
     "archived": false,
     "on_spotify": true
-  },
-  {
-    "id": 429,
-    "genre": {
-      "id": 411,
-      "genre": "a",
-      "items": [
-        {
-          "item_id": 828,
-          "item_class": "MusicAlbum"
-        },
-        {
-          "item_id": 429,
-          "item_class": "MusicAlbum"
-        },
-        {
-          "item_id": 335,
-          "item_class": "MusicAlbum"
-        }
-      ]
-    },
-    "author": "g",
-    "label": "f",
-    "publish_date": "2022-02-02",
-    "archived": false,
-    "on_spotify": false
-  },
-  {
-    "id": 335,
-    "genre": {
-      "id": 411,
-      "genre": "a",
-      "items": [
-        {
-          "item_id": 828,
-          "item_class": "MusicAlbum"
-        },
-        {
-          "item_id": 429,
-          "item_class": "MusicAlbum"
-        },
-        {
-          "item_id": 335,
-          "item_class": "MusicAlbum"
-        }
-      ]
-    },
-    "author": "yh",
-    "label": "ter",
-    "publish_date": "2022-02-02",
-    "archived": false,
-    "on_spotify": false
   }
 ]
