+++ src/app.rb
@@ -1,10 +1,9 @@
+require './src/classes/music_album/manage_music_album'
 require './data/persistors/genre_persistor'
 require './data/persistors/music_album_persistor'
+require './src/classes/book/book'
 require './data/persistors/book_persistor'
 require './data/persistors/label_persistor'
-require './data/persistors/item_persistor'
-require './src/classes/book'
-require './src/classes/music_album'
 
 class App
   def initialize
@@ -25,7 +24,6 @@ class App
       '8' => method(:add_music_album),
       '9' => method(:add_game)
     }
-    ItemPersistor.deserialize_items_all(@genres, @music_albums, @books, @games)
   end
 
   def run
@@ -69,7 +67,6 @@ class App
     @genres.each_with_index do |genre, index|
       puts "(#{index + 1}) - #{genre.genre}"
     end
-    puts @genres.inspect
   end
 
   def list_all_labels
@@ -87,7 +84,7 @@ class App
   end
 
   def add_music_album
-    MusicAlbum.add_music_album(@music_albums, @genres)
+    ManageMusicAlbum.add_music_album(@music_albums, @genres)
   end
 
   def add_game
