+++ src/classes/author/author.rb
