+++ src/classes/book/label.rb
@@ -1,4 +1,4 @@
-require_relative './item'
+require_relative '../item/item'
 
 class Label
   attr_accessor :title, :color
