+++ src/classes/genre/genre.rb
@@ -1,0 +1,14 @@
+class Genre
+  attr_accessor :items, :id, :genre
+
+  def initialize(genre)
+    @id = Random.rand(1000)
+    @genre = genre
+    @items = []
+  end
+
+  def add_item(item)
+    @items << item unless @items.include?(item)
+    item.genre = self
+  end
+end
